replace source with vanilla libiconv-1.15

+++ source/lib/hz.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 1999-2001, 2008 Free Software Foundation, Inc.
+ * Copyright (C) 1999-2001, 2008, 2016 Free Software Foundation, Inc.
  * This file is part of the GNU LIBICONV Library.
  *
  * The GNU LIBICONV Library is free software; you can redistribute it
@@ -14,7 +14,7 @@
  *
  * You should have received a copy of the GNU Library General Public
  * License along with the GNU LIBICONV Library; see the file COPYING.LIB.
- * If not, write to the Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ * If not, see <http://www.gnu.org/licenses/>.
  */
 
 /*
@@ -28,7 +28,7 @@
  */
 
 static int
-hz_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
+hz_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, size_t n)
 {
   state_t state = conv->istate;
   unsigned int count = 0;
@@ -36,7 +36,7 @@ hz_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
   for (;;) {
     c = *s;
     if (c == '~') {
-      if ((unsigned int)n < count+2)
+      if (n < count+2)
         goto none;
       c = s[1];
       if (state == 0) {
@@ -48,13 +48,13 @@ hz_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
         if (c == '{') {
           state = 1;
           s += 2; count += 2;
-          if ((unsigned int)n < count+1)
+          if (n < count+1)
             goto none;
           continue;
         }
         if (c == '\n') {
           s += 2; count += 2;
-          if ((unsigned int)n < count+1)
+          if (n < count+1)
             goto none;
           continue;
         }
@@ -62,7 +62,7 @@ hz_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
         if (c == '}') {
           state = 0;
           s += 2; count += 2;
-          if ((unsigned int)n < count+1)
+          if (n < count+1)
             goto none;
           continue;
         }
@@ -77,7 +77,7 @@ hz_mbtowc (conv_t conv, ucs4_t *pwc, const unsigned char *s, int n)
     return count+1;
   } else {
     int ret;
-    if ((unsigned int)n < count+2)
+    if (n < count+2)
       goto none;
     ret = gb2312_mbtowc(conv,pwc,s,2);
     if (ret == RET_ILSEQ)
@@ -97,7 +97,7 @@ ilseq:
 }
 
 static int
-hz_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
+hz_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, size_t n)
 {
   state_t state = conv->ostate;
   unsigned char buf[2];
@@ -148,7 +148,7 @@ hz_wctomb (conv_t conv, unsigned char *r, ucs4_t wc, int n)
 }
 
 static int
-hz_reset (conv_t conv, unsigned char *r, int n)
+hz_reset (conv_t conv, unsigned char *r, size_t n)
 {
   state_t state = conv->ostate;
   if (state) {
